[NFC] Avoid repeated work in Precompute of GC allocations (#7763)
When we precompute struct.new, we have a cache of allocations, as we
need the ID of allocations to remain the same. The logic for managing that
cache was actually first creating a new instance, then throwing it away if
we used the cached value... so we ended up visiting the children and
computing them etc., with exponential (wasted) work in the worst case.

To fix that, just check the cache first.

Fixes #7760