[Custom Descriptors] Fix TypeMerging topological sort (#7764)
TypeMerging uses a topological sort to process supertypes first in
various places. It also treats descriptor chains as single units
represented by their base described type. Types in WebAssembly cannot
have more than one supertype, but the types in a descriptor chain will
have different supertypes and those supertypes may not necessarily be
all in a single descriptor chain together. As a result, when descriptor
chains are treated as units, it is possible for one chain to have
multiple supertype chains. The topological sort in TypeMerging did not
previously take this into account, and as a result the pass tried to
merge types in the wrong order, leading to assertion failures. Fix the
topological sort to correctly treat descriptor chains as units.

Fixes #7757.