[Custom Descriptors] Fix AbstractTypeRefining for exact casts (#7768)
AbstractTypeRefining would previously optimize a cast to
`(ref (exact $uninstantiated))` to `(ref (exact $instantiated))` when
traps never happen and `$instantiated <: $uninstantiated`. This is not
correct, however, because it transforms a cast that will never succeed
to a cast that might succeed.

Generalize our existing logic for fixing up descriptor casts to also fix
up exact casts. Move the fixups to occur before the types are rewritten
so we don't lose track of the original types of the casts.