[Custom Descriptors] Consider desc effects in HeapStoreOptimization
HeapStoreOptimization moves subsequent sets back into initial
allocations, so it must ensure the set values can be reordered before
all of the original set values that it will precede. If the allocation
has a descriptor operand, then the set value must be reordered before
that as well, but we were not checking that operand for effects. Fix it.