Fix SubtypingDiscoverer with shared i31 (#7774)
SubtypingDiscoverer previously noted that the child of I31Get
expressions were subtypes of i31, but they can alternatively be subtypes
of (shared i31) and this was not handled. This should properly be
modeled as a new kind of type constraint in SubtypingDiscoverer, but the
more robust solution would be to refactor SubtypingDiscoverer to use
principal types so the sharedness can be a type variable. For now, just
look at the child to figure out what shareness to use.

This bug was found because it caused an infinite loop in Unsubtyping, but
add assertions that would have caught it directly and add a simpler test
that would have triggered the assertions had they existed before.

As drive-bys, add a new `insert` overload to InsertOrderedSet to make
the compiler happy with UNSUBTYPING_DEBUG set and make the whitespace in
another unsubtyping test consistent with the whitespace in the rest of
the file.

Fixes #7773.