[wasm-split] Do not create duplicate trampolines
When a function is exported and also referenced somewhere, like in an
`elem` section or in a `ref.func` instruction, we create a trampoline
for it twice, once for its export in `thunkExportedSecondaryFunctions`
and again for its reference in `thunkExportedSecondaryFunctions`. This
makes it to be generated only once in the case.