[NFC-ish] Refactor away callTable (#7776)
Historically, wasm only had call_indirect, and we put callTable as a
method on the interpreter's ExternalInterface. Later, wasm added
table.get etc., and we added tableLoad to that interface. Given that,
we don't really need callTable: It is simpler for e.g. visitCallIndirect
to use tableLoad + do a call of the loaded reference.

This removes a bunch of duplicate code around the codebase
(implementers of callTable), and makes working on visitCallIndirect
simpler (which I will need for stack switching).

Some minor test changes happen here, mostly that the trap
message is slightly different in some cases. The one interesting
change, which makes this not NFC, is that a TODO is now
complete in test/lit/ctor-eval/return_call.wast