[Custom Descriptors] Handle traps in instantiation in Merge fuzz handler
With custom descriptors, it's possible for instantiation to trap when
a `struct.new` in a global initialiizer receives a null descriptor.
Update the Merge fuzz handler so that if the randomly generated second
module traps in instantion, it checks that the merged module also traps
in instantiation.

As a drive-by, remove some unnecessary feature flags passed to
wasm-merge and wasm-opt.