[wasm-split] Support --no-placeholders option in multi-split
This supports `--no-placeholders` option with `--multi-split`. Currently
`--multi-split` does not accept `--no-placeholders` option and set it to
true by default internally.

Because `--no-placeholders` is false by default and this has been the
case for the default two-way split, this means this PR is a breaking
change for multi-split users.
https://github.com/WebAssembly/binaryen/blob/8ad26cbaca2a3a8982789b25d9a279dc26e85f15/src/tools/wasm-split/split-options.cpp#L210-L218

But it will be inconsistent if we set the option by default to true in
case of two-way split and false in case of multi-split.

About the test:
- This changes `CHECK-A`, `CHECK-B`, and `CHECK-C` to `MOD1`, `MOD2`,
  and `MOD3`
- Because previously `--no-placeholders` was the default, the
  expectation lines for `PRIMARY` didn't have any placeholders. Now it
  does.
- `***-OPTIONS` check lines now test both options, `--import-namespace`
  and `--no-placeholders`.