[Custom Descriptors] Validate features in TypeBuilder
The Binaryen validator does not validate all types because it does not
want to pay the cost of walking the module to gather the types.
Furthermore, even if it did walk the module to gather types, it would
still have no way of validating completely unused types. However,
correctly validating that a module's types use only the allowed features
can be important, for example to prevent the fuzzer from using initial
contents that use a feature with a handler that cannot handle that
feature. To achieve this validation without paying the cost in the
validator, validate features used by types in TypeBuilder instead. For
now the only feature validated is custom descriptors.

To avoid introducing new errors, ensure that all features are available
when parsing in wasm-dis and wasm-reduce. Also tweak error messages to
make the binary and text parsers consistent so we can simplify the test.