[wasm-split] Support --placeholdermap for --multi-split
This adds support for `--placeholdermap` for `--multi-split`. In the
normal two-way split mode, `--placeholdermap` emits a single
placeholdermap file for the primary module, but semantically a
placeholdermap should exist per split module.

I guess we should switch to make the placeholder file name follow the
secondary module for the two-way split for consistency?