Fix determinism of ClusterFuzz fuzz handler (#7788)
The first time the Clusterfuzz handler runs, it does a lot of setup work
that can be annoying when trying to run the fuzzer briefly to smoke out
very obvious problems. To avoid that annoyance, we do not run the
Clusterfuzz handler until we have already fuzzed for at least 30
seconds. However, we used to control this in the `can_run_on_wasm`
method, which caused reproducibility issues because the number of calls
to `random` functions would differ before and after that 30-second
cutoff. Specifically, `relevant_handlers` would have a different size
before and after that cutoff, so the number of `random` calls to compute
`filtered_handlers` would be different.

To avoid these reproducibility problems, simply return early from the
`handle` method of the Clusterfuzz handler if it has not yet been 30
seconds.