[Stack Switching] Update continuation types based on resume tag & target block (#7786)
The resume tag defines how the continuation should be called,
perhaps with different parameters than before. To handle that,
update its type and also properly handle function parameters
(if we are resuming with a different type, we should not error -
the function params are already saved as local state, and we
just need to reload them). The actual type also depends on the
target block.

This allows the next part of the spec test to run.