Support cross-module calls in the interpreter (#7787)
Previously a runtime instance of a function was just the function name.
That is fine for optimizations, and even for executing a single module,
but not cross-module calls.

To support that, add a FuncData (parallel to ExnData, ContData, etc.). The
constructor now requires that (but keep makeFunc as a simple helper
for cases when we just want the name).

A FuncData contains info about the module instance the function is in,
and a std::function to actually call it. We create a proper thunk when we
create a function literal, and in call_ref, we use it to do the actual call.