[wasm-split] Add -all to wasm-dis RUN lines
In tests that we use `-all` options to `wasm-split`, we need to add
`-all` to `wasm-dis` command lines as well to get the correct results.
This can generate incorrect disassembly because, when we give `-all` to
`wasm-split` we generate multiple tables, but if `-all` is missing in
`wasm-dis`, tables will be missing in `call_indirect`s, making them
always refer to the first table.
https://github.com/WebAssembly/binaryen/blob/21464e7533ef235634c8d62442a8a10cafff6568/src/passes/Print.cpp#L506-L509