[Stack Switching] Implement the simple case of cont.bind (#7793)
In the simple case, all parameters are bound. This is common in the
spec test, and will unblock a bunch of stuff there.

To do this, add full support for resuming without a resume expression,
that is, without a Suspend that we need to unwind back to. Specifically,
when we resume a bound continuation, the only thing we need to do
when resuming is to apply the bound arguments, and then
continue execution normally.