[Stack Switching] Handle imported tags properly (#7801)
We were using Tag names, but the internal name in a module may
overlap with other modules etc. - we really need to look at the
actual imported Tag, if it is imported. To do that, store Tag* and
add a way to get the canonical Tag* - i.e., the imported one if
imported, so all references, including from other modules, all
refer to the same thing.

To be able to test this, make `callImport` return a Flow (so we can
suspend through import calls).

As a bonus, replacing `Name tag` with `Tag* tag` makes our data
structures smaller.