[Stack Switching] Handle nested continuations (#7803)
The key fix here is that we don't know what the continuation
function (the thing to call, to continue onwards) is, at the time of
suspending. Rather, that function is called from a resume, to
resume execution from there, so do so when we handle a
suspend in a Resume.

And, when we resume "through" a suspend (due to a nested
suspend), then we should just rewind the call stack.

With this, the next big spec testcase passes.

Also fix a minor issue with trying to put breaking values on
the value stack - only value stack values go there. The new
testcase enabled here trips on that.

Also improve some debug logging.