[NFC] Update ChildTyper callback for unknown constaints (#7814)
When ChildTyper does not have enough type information to generate type
constraints (for example, because a child whose type provides crucial
information is unreachable), it previously allowed the user to determine
whether to return early or not. But not returning early would then lead
to an assertion failure or other bad behavior, so it is never a
reasonable choice. The only real choice a user has is to return early
and ignore the situation (which is what TypeSSA wants) or to fail loudly
because this should never happen (which is what IRBuilder wants).
Simplify the callback to allow customizing this behavior but not whether
the early return ocurrs.