[Stack Switching] Implement resume_throw in the interpreter (#7815)
Pretty straightforward: Add an exception tag to ContData, and throw if
it is set.

With this, the next big spec test passes.

Throwing from the first resumption is not in the spec test, so add that
manually (for us, but I guess not the spec interpreter, this checks
another code path).

This uncovered an existing bug with StackValueNoter. We added one of
these during Resume, to ensure a stack value noting scope always
existed. But we track those by ModuleRunner, so it doesn't help when we
do a cross-module suspend. Just remove that extra logic, and handle the
lack of a scope directly, which makes this all a lot simpler.