[Stack Switching] Ignore attempts to stash resume info for suspends without a continuation (#7819)
When there is no continuation (we are not running in one), then a
suspend will lead to a trap on the host side, of "unhandled
suspend". We do need to make sure not to error on the way, so
any attempt to save resume info must be skipped.

Spec tests did not catch this because they did not have an
enclosing block, and the block tries to save resume info.