[Strings] Be stricter when checking builtin types (#7829)
String builtins should not only have the expected signatures, they
should also have the expected types. For example, if a builtin has a
signature that is in a nontrivial rec group or has a supertype or is not
final, that should be a validation error. Reject such incorrect types in
StringLifting.