Fuzzer: Improve handling of small inputs and their debugging (#7832)
Add a new env var to control the fuzzer input size,

BINARYEN_FUZZER_MAX_BYTES

When set, we will truncate the random bytes we receive as
input in translate-to-fuzz. This makes it simple to bisect on the
input, for example.

Also make the fuzzer notice that the input has run out in more
cases, and stop emitting things. This makes small inputs to
the fuzzer smaller in output.