Fix regression from #7822 about exnrefs with nulls and unknowns (#7835)
RootLocation, added in #7822, was overly-ambitious in how much it
simplified. It turns out that there are situations where we need the same
type for both a null and an "unknown" of that type, and a single RootLocation
is therefore not enough (the specific situation the fuzzer found involves
exnref, but perhaps others are possible too).

To fix this, refactor back to two locations, NullLocation and TypeLocation,
where the latter is at least shared by exn and cont as a place with an
unknown value of a type.