[wasm-dis] Restore features section in output
This restores the features enabled by the features section (if it
exists) right before the printing time, to make the features section
comment consistent with the input file's features section. If the input
binary does not have the features section, it won't print anything.

`ModuleReader::getFeaturesSectionFeatures` was adopted from @tlively's
commit:
https://github.com/WebAssembly/binaryen/commit/ed36a8cd8aa46219b054f736f0fe3bf963a4f763

Because we don't set the features to `Features::All` at the printing
time, `(type $..)` in `(func)` definitions in the tests have been
removed.

Fixes #7836.