Always print tables for call_indirects
This unconditionally enables printing of the table reference in
`call_indirect`s, because it can cause incorrect wast generation if the
feature section/flags were not provided. And our parser can parse it
without any feature flags anyway. See #7802 for the details on this.

This is a preparation for restoring the features section at the time of
printing for wasm-dis (to fix #7836), but because this causes many lines
of test changes, I'm uploading this as a separate PR. The test changes
are because now we print the table reference in wasm-opt results as
well, not only in wasm-dis results.