[Custom Descriptors] Fix TypeMerging sibling criteria (#7842)
When merging sibling types, we must first partition the types by their
supertypes and shapes. With custom descriptors, we consider full
descriptor chains as single units to be merged. We had updated the shape
partitioning to account for this, but not the supertype partitioning. As
a result, we were incorrectly merging chains with different descriptor
supertypes as long as the base described types had no supertype.

Fix the bug by partitioning not just on the base described type's
supertype, but the sequence of all supertypes in the chain.