Print tables for call_indirects for nonzero table index (#7843)
Rather than always printing the table like in #7839, this prints the
table only when either reference types is enabled or the table index is
not 0 (even if the features section or `--enable-reference-types` is not
provided)

This addresses the concern in #7802, while not printing table when it is
not necessary.

There is no test changes because
- At the moment wasm-dis enables all features so wasm-dis always prints
tables in `call_indirect`s. This will change in #7840.
- wasm-opt results didn't print table so far when reference-types is not
enabled, and all `call_indirect`s in those tests have table index 0.