Flatten: add relaxed mode supporting try_table
This introduces a "relaxed" mode to the flatten pass, which allows it to process try_table expressions. In this mode we preserve blocks return values if those blocks are used as catch destinations and we also preserve breaks with values if they target these blocks. To make this useful for asyncify, blocks return values are still saved into locals.