Asyncify: add basic support for try_table
This uses the new "relaxed" flat IR which supports try_table. To make it work with asyncify we add support for the new flat IR "local.set with a block" expression, where we need to add a dummy local.get at the end to make the catch block reachable when rewinding.