wasm-merge: Avoid quadratic time in validation [NFC] (#7845)
The actual merging logic in wasm-merge is careful to run in linear time
(since #5709). However, validation could actually be quadratic, since we do
it after each module that we merge in. Since we validate by default, this
made wasm-merge very slow sometimes.

To fix that, validate by default once at the very end. When in pass-debug
mode, which validates after each pass, validate after each module (so we
keep the option for precise debugging of problems).

Also, move some validation code being the validation flag, so that
--no-validation is even faster.

On a large testcase I am investigating (350 modules, 41 MB wasm after
merge), this makes us go from 5 minutes down to just 19 seconds. With
--no-validation, we drop to 17. (That final number is about how long it
takes to run wasm-opt 350 times to round-trip each of those modules,
so it is shows wasm-merge's merging is very fast.)