TypeUpdating: Fix heap type updates on heap type children (#7848)
`GlobalTypeRewriter::getTempType` had the right logic to handle a
Type, which included the right logic to handle the HeapType of a
Type: Look in `typeIndices` and use that mapping if it is present. We
need the same thing for HeapTypes as well, which this adds in a
new `getTempHeapType`.

This was not noticed before because Continuations are the first
thing to have a direct HeapType child which is not through a Type.
So Types worked before, but not direct HeapType children, which
ended up mapped to the old type before the rewriting.

Changes to `wasm-type.h` are NFC but I think make it easier to read,
as it makes that logic entirely parallel to the lines above it.

One existing test has an update, which fixes it by removing the
ref to the old type.