[GC] Precompute: Fix ref.eq comparisons of structs with nested effects (#7857)
In the testcase here, the nested tee make the outer struct not precomputable
when we care about effects (tee is an effect). But we could precompute it
otherwise, and that ended up causing confusion, with no caching being done
of the canonical allocation for it, meaning each time we scanned it we saw
another struct, different in identity.

To fix this, implement the TODO to cache in both modes. I thought that would
be useful only for speed, but turns out it is necessary for correctness. Luckily
doing so is quite simple.