[wasm-dis] Restore features section in output (#7840)
This effectively restores pre-#7802 status, except for the case the
input binary has the features section. If the binary does not have it,
we restore the default features (mutable-global + sign-ext) + whatever
is given in the command line before the printing. If the binary has the
features section, we just restore it before the printing.

This currently does not handle the case when there is the features
section and the user gives more features flags in the command line, but
in the current interface it is cumbersome to handle. And we were not
handling that anyway pre-#7802, so it is not a regression.

And we preserve the `FeaturesSet::All` setting at parsing time so
@tlively's
[concern](https://github.com/WebAssembly/binaryen/issues/7836#issuecomment-3198234378)
will not be an issue.

Because we preserve the features section, #7836 will be addressed too.

The main motivation for #7802 was `call_indirect` could be incorrectly
printed, and that was also separately addressed by #7843.

`ModuleReader::getFeaturesSectionFeatures` was adopted from @tlively's
commit:

https://github.com/WebAssembly/binaryen/commit/ed36a8cd8aa46219b054f736f0fe3bf963a4f763

Because we don't set the features to `Features::All` at the printing
time, `(type $..)` in `(func)` definitions and `(table ..)` in
`call_indirect`s have been removed.

Fixes #7836.