[Stack Switching] Fix resuming of return calls (#7860)
A return call is a form of control flow, so we need to handle it. Specifically,
the original function is unwound before calling the next one, so when we
resume, we should just call the next one. To do that, save the call target
along with the locals during unwinding.