[Custom Descriptors] Optimize away unneeded descriptors in GlobalTypeOptimization (#7872)
This adds descriptor support to struct-utils.h as an extra field. It is
reported as index "-1" when an index is used in the API, allowing
passes to skip it where irrelevant.

That new support is then used in a single pass, GTO, where we just
see if a descriptor can be removed, and remove it if so.