Add a TypeOrdering pass
The binary writer already takes care to order rec groups to minimize the
size taken to encode type indices throughout the module, but it cannot
order types within a rec group because that would change their identity.
MinimizeRecGroups can be used to split the type section into small
recursion groups that can be ordered by the binary writer, but that pass
does not optimize the order of types within a rec group. Furthermore, if
the minimal early rec groups are large, all the extra types taking up
valuable index space can cause a large code size overhead.

To get the best possible outcome, add a TypeOrdering pass that creates a
single large rec group, giving the types the most possible flexibility
to be moved around, and orders the types within it. As we do in
ReorderGlobals, try multiple tolological sorts with different factors
for adjusting weights based on successor types, then choose the best
ordering. Refactoring so that ReorderGlobals and ReorderTypes can share
more code is left as future work.

To avoid reimplementing all of the GlobalTypeRewriter utility, which
already rewrites all the private types into a single rec group, factor
out just the part that sorts the types and make it overrideable in
subclasses.