[Stack Switching] wasm-ctor-eval: Properly handle continuation serialization errors in results (#7883)
#7877 gave up when we couldn't serialize things, in various situations. But
one of the cases, when we can't serialize the results, we checked AFTER we
already "commit" the changes to the module. That is, we need to check for
errors and stop if we find any, and only then apply execution results to the
module, and we added a check that was after the application.

To fix this, we need to be careful and know that if we will care about the
serialization of the results then we cannot execute continuations, as they
may end up needing to be serialized. Doing anything more precise would
be significantly more complex.