[Custom Descriptors] Fix CFP on an exact ref.get_desc (#7886)
`ref.get_desc` inherits its input ref's exactness: When called on an
exact ref, it returns the exact descriptor type. CFP was not aware of
that, and thought any subtype can appear as well.

That was not only a missed optimization, but also a validation bug:
`ref.get_desc` returned an exact type, but if we consider values
from subtypes, we may emit a select over two values that is no
longer exact.

This also helps optimize `struct.get`, as the logic is shared, so now
`struct.get` of an exact type (and immutable field) will ignore
subtypes and look only at data from `struct.new`.