[wasm-split] Support table64
This makes wasm using table64 (included in memory64) can be used in
wasm-split. To be strict we should use different data types than
`Index`, which is `uint32_t`; for memory64 we use `Address` class to
represent the address. But it is unlikely that table indices would ever
go above `uint32_t` range, and we are not even sure if engines support
that many. To be strict we should be using `Address` for table indices
as well, but the use of `Index` for table indices is currently not only
limited to module-splitting utils, so I think we can cross that bridge
when it becomes really necessary.

Fixes #7716.