Honor CPU affinity in ThreadPool::getNumCores
This reports that number of CPUs that are actually usable by the current
process.

This means that I can do something like `taskset -c 0 emcc hello.c -O2`
and the internal call to `wasm-opt` within emcc will only use a single
core.

Ideally this would work on macOS and windows too, but I'm not even sure
how easy it is to control affinity on those OSes.