[GC] Use cfp-reftest pass (#7905)
This pass has been used by j2wasm for a while, and is useful there, so
enable it for general use.

This pass can grow code size in some cases, if it emits a ref.test that is not
later optimized out. For that reason only enable it in -O3+, that is, when
optimizing aggressively for size.

Also fuzz this pass more.