[wasm-split] Support table64 (#7904)
This makes wasm using table64 (included in memory64) can be used in
wasm-split.

To be strict we should use different data types than `Index`, which is
`uint32_t`; for memory64 we use `Address` class to represent the
address. But it is unlikely that table indices would ever go above
`uint32_t` range, and we are not even sure if engines support that many.
We can consider using `Address` for table indices, but the use of
`Index` for table indices is currently not only limited to
module-splitting utils, so I think we can cross that bridge when it
becomes really necessary.

Note that when creating a new table, we don't create it as an i64 table
even if memory64 is enabled, because it is not necessary. We only deal
with existing i64 tables.

Fixes #7716.