[Custom Descriptors] Fix shared GTO placeholders
When we optimize a descriptor to describe a placeholder type because the
original described type no longer needs a descriptor, we were not
previously ensuring that the sharedness of the placeholder matched the
sharedness of the descriptor type. Make sure the shareness matches to
avoid validation errors when rebuilding the types.