[Custom Descriptors] Fix GTO placeholders for unused types (#7908)
GTO emits placeholder describee types when it optimizes descriptor types
that must remain descriptors. However, it previously tried to emit a
placeholder describee type even for a descriptor type that is unused in
the module and therefore not included in the rebuilt types. These
dangling describees caused validation faillures when rebuilding the
types. Fix the problem by emitting placeholder describees only for
descriptors that will be rebuilt.